pyproject.toml
[project]
name = "guardian"
version = "0.1.0"
description = "Unified monitoring for npm packages, GitHub repos, and Fly.io/Vercel deployments"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pygithub>=2.2.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "slowapi>=0.1.9",
    "aiosmtplib>=3.0.0",
    "prometheus-client>=0.20.0",
    "pyyaml>=6.0.0",
    "mcp>=0.1.0",
]
[project.scripts]
guardian = "guardian.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "hypothesis>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

